Refactor tasks.py to use functions and add argparse
Basically rewrote the whole program to use functions.

Also added argparse functionality to allow user run the mouse position checker
without having to go in and uncomment/comment different lines of code. As it
stands, you do still have to manually change the link location tuples in the
code. I couldn't be bothered to implement them being changed by the user as the
script is running.